+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Upcoming Version
+
+**Added:**
+- Azure SQL Service - Support for SQL Server, SQL Databases and Data warehouse added.
+
 ## 0.0.8 
 
 **Added:**